Restarting Game stays on the same page
Changed a bunch of code around such that clicking the "restart" symbol will now create a new game WITHOUT restarting the entire page.

Doing this revealed a bunch of bad code:
* There are a lot of repetitive functions.  * Game Logic, SVG Interactions, and graphics are not well-encapsulated.
* The codebase is not well organized.

Try using some of the object-method models that were in the first version (the one using Canvas).  It cleared up the game logic, and was much more separated from the way graphics were handled.